requirements.txt
Django>=5.0,<7.0
djangorestframework>=3.15
gunicorn>=21.2
whitenoise>=6.6
